refractor: change notifications to controller class
